Word search game.
The computer chooses a random word from a module of more than 30,000 words. Then it says the number of letters in the word and asks us to find this word by letter. After each attempt, it shows the letters found in the word and their place in this word, as well as a list of the letters we have chosen. If the letter is repeated, it notifies us about it. After the end of the game, it shows the number of our attempts and congratulates on the victory.